Add abstraction method to lic. purpose points (#2642)
https://eaflood.atlassian.net/browse/WATER-4645

> Part of the work to migrate return versions from NALD to WRLS

We recently [Added new points and sources tables & amend existing](https://github.com/DEFRA/water-abstraction-service/pull/2638) to the service. This not only supports our work on return versions, it means we can move away from abstracting point information from the licence JSON blob in `permit.licence`.

We're also looking to migrate the pages linked to from the view licence summary, which are currently in the legacy [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui). One of them is information on points. We thought we'd captured everything 'point' related in the change but have since spotted 'Means of abstraction'.

This is linked to an abstraction purpose point in NALD and gives how water will be abstracted there, for example, "Surface Mounted Pump (Fixed)". We must capture this information as we import a licence's point information from NALD. So, the first step is to update the `water.licence_version_purpose_points` table to add a new `method` field that will allow us to store this information.