Alter fix script for miss-tagged credited trans.
https://eaflood.atlassian.net/browse/WATER-4710

In the last change we added a [Data fix migration for miss-flagged transactions](https://github.com/DEFRA/water-abstraction-service/pull/2644). In the legacy presroc supplementary billing engine when a debit is reversed and used as a credit it adds a flagged to the debit: `is_credited_back`. It also sets `source_transaction_id` on the new transaction so it is possible to see the link between the two.

What we found was that the PRESROC supplementary when it fails can mess up these flags. We found

- transactions with `is_credited_back` set to true, even though their IDs do not appear in any `source_transaction_id`
- transactions with a populated `source_transaction_id`, but the source transaction does not have its `is_credited_back` flag set to true

The script we added was supposed to fix the the state of the `is_credited_back` flag on these transactions, and the correct bills would then be generated for the licences reported to us by Billing & Data.

However, this was not the case. One licence where we were not expecting a bill was still generating one. After working through all the transactions and matching up historic to new transactions we found our fix for the second issue had made in an invalid assumption.

We assumed it was always a historic debit being reversed as a credit. In this case however, it was a credit that was reversed as a debit.

So, in this change we're going to alter the original migration, and move the second part to a new migration. This avoids us having to manually correct any environment where the migration has already been run.

Then in the new migration we'll remove our where clause `AND bt2.is_credit = TRUE` from the update query. This means _all_ transactions with a populated `source_transaction_id``, but the source transaction does not have its `is_credited_back` flag set to true will be corrected. We've also confirmed in testing this resolves the issue for the licence (no bill is raised when the supplementary is run).