Alter fix script for miss-tagged credited trans. (#2645)
https://eaflood.atlassian.net/browse/WATER-4710

In the last change, we added a [Data fix migration for miss-flagged transactions](https://github.com/DEFRA/water-abstraction-service/pull/2644). In the legacy PRESROC supplementary billing engine, a flag is added to the debit: `is_credited_back` when a debit is reversed and used as a credit. It also sets `source_transaction_id` on the new transaction so it is possible to see the link between the two.

We found that the PRESROC supplementary can mess up these flags when it fails. We found

- transactions with `is_credited_back` set to true, even though their IDs do not appear in any `source_transaction_id`
- transactions with a populated `source_transaction_id`, but the source transaction does not have its `is_credited_back` flag set to true

The script we added was supposed to fix the state of the `is_credited_back` flag on these transactions, and then the correct bills would be generated for the licences reported to us by Billing & Data.

However, this was not the case. One licence for which we were not expecting a bill was still generating one. After working through all the transactions and matching up historic to new transactions, we found our fix for the second issue had been based on an invalid assumption.

We assumed it was always a historic debit reversed as a credit. In this case, however, it was a credit reversed as a debit.

In this change, we're going to alter the original migration and move the second part to a new migration. This avoids manually correcting any environment where the migration has already been run.

In the new migration, we'll remove our where clause `AND bt2.is_credit = TRUE` from the update query. This means _all_ transactions with a populated `source_transaction_id` but the source transaction does not have its `is_credited_back` flag set to true will be corrected. We've also confirmed in testing that this resolves the issue for the licence (no bill is raised when the supplementary is run).