Spike fix pdf generation issue
The pdf generator uses puppeteer to create a pdf.

We have an issue in our local environments where the pdf generations is failing. The issue can be found here - https://github.com/DEFRA/water-abstraction-team/issues/102.

This change proves the fix works. But does not account for our deployments and so should not be merged.

We can look for a solution to support local and a deployed state,