Move 2PT review result tables into 'water' schema
https://eaflood.atlassian.net/browse/WATER-4739

The work to build the two-part tariff match & allocation engine inspired our solution to having to work with 'models' split across multiple schema. By adding a view to the public schema that referenced a table elsewhere, we could pretend that the project is only having to deal with a single database. We could start creating relationships in our models for tables that sat in different schemas!

We expanded the scope of this solution to also help us clean-up the tables we were working with. When we create the views, we rename columns to use standard naming conventions. We rename columns to be consistent across the views. We also drop columns that have become redundant, those that were built some some future 'need' that was never needed!

That fed another idea: we could slowly migrate all data elements into the 'public' schema (PostgreSQL's default and what would be used if the previous team hadn't opted to use schemas. Overtime, 'public' would be the cleansed version of the service database. So, any new table we needed we created in the `public` schema.

Well, that was the plan. Since then we've worked on other features which have required new tables. Unlike two-part tariff though, these tables have been extensions of existing domains. For example, our return versions setup work needed a new table to link return requirements to points. So, `water` already has `return_versions`, `return_requirements`, and `return_requirement_purposes`. It didn't make sense to put `return_requirement_points` in a completely different schema.

In fact, our experience with the service has shown us the `water` schema is the de facto schema. It came after the others and is in large part a 'first draft' at tidying up and normalising the other schemas.

Today, putting a few tables in `public` now feels like we are complicating matters by adding yet another schema. The views have also had to change to incorporate fields we wouldn't have built but are needed to support working with the legacy functionality. The idea of slowly migrating everything to `public` no longer seems feasible. Instead, we believe we first need to complete the work of making all schemas bar `water` defunct. Then, clean it up till it is in a state we are happy with. At which point we can choose to rename it `public` and we will have achieved our original vision.

So, before two-part tariff goes live and the review results tables become used for real, we're moving them into the `water` schema. This change takes care of creating the new tables in the 'water' schema and copy any existing data across.