Fix xml parser prior to upgrading to node v22 (#2656)
https://eaflood.atlassian.net/browse/WATER-4729

We want to upgrade to node v22, but we have a blocker with the xml parser we are using.

The tool being used is libxmljs and has some open issues:

- https://github.com/libxmljs/libxmljs/issues/664
- https://github.com/libxmljs/libxmljs/issues/660

Mainly the issues are:

- is it being maintained
- is there going to be a release to make it work with node v22.

Rebuilding the package locally during our npm install has not worked, (through minimum effort) and if we did make it to build there is no guarantee it will actually work when built.

This spike investigates the effort required to get any xml parser to work with the least possible effort.

The change suggest using [libxmljs2](https://github.com/marudor/libxmljs2). Which works.

However, if you look at the documentation it states : "LibXML bindings for node.js This package was forked as the original one is fairly unmaintained."

Sounds good. But now it clearly now states it is "NO LONGER MAINTAINED".