Add a constraint to the table
This will support the event that the check is performed a second time before the WRLS record has been updated, or the end date change process has had a chance to run.

If the `/licences/end-date/check` process flags the same licence again, for the same date type, we can use the constraint to apply an Upsert query and update the dates to the latest information.

This avoids us erroring if we attempt to insert the record for a second time.