Doh! Add change_date field
Forgot to add this when creating the table. Our code in water-abstraction-system, having identified what date change needs to be captured, also determines what change date to use, for example, if 2 dates have been updated at the same time.

We need this to be able to process the change in our later process.