Fix charge versions 2pt flagging mechanism
https://eaflood.atlassian.net/browse/WATER-4951

Recent work has been completed to consolidate code where licences should be flagged for pre-sroc, sroc and two-part tariff supplementary billing. In the new flagging process a bug has been raised. The bug specifically deals with changes on a charge version. If a charge version was 2pt and is being changed to be a non-2pt charge version then the flagging logic is incorrect and the licence isn't being flagged when it should be. Equally, the flagging logic doesn't consider non-chargeable charge versions as well. A PR in system fixes the incorrect flagging issue whilst this PR accommodates new params being passed to the flagging service.