Update notification messageStatuses to include pending
https://eaflood.atlassian.net/browse/WATER-4961

We had an assumption that using a status of 'sending' for a scheduled notification did not conflict with the existing legacy code.

This is not the case.

To avoid interacting with the legacy notification status update mechanism we have introduced a new status for scheduled notifications which is 'pending'.

This is, for all intents and purposes, the same as sending (without the legacy code attempting to update the status).

The legacy code report will be updated to allow this new 'pending' status.

We also previously added a feature toggle to turn off the whole batch notification jobs when instead we should have only toggle the notification check status.

This change update the toggle to only 'turn off' the check status logic.