Remove charge categories sync job
There is a CSV file (`billing-metadata/charge-categories.csv``) held in the `wabs3-maintenance`` bucket which contains a list of charge categories.

Every 6 hours a BullMQ job checks for changes and either inserts or updates the changes.

But it is pointless!

There has never been a change in the file since it was first dropped. Also, because it sits in the `production` bucket we couldn't upload a different one without a Normal RfC. This means we'd need to wait 10 business days for the changes to appear.

If the data is managed using a seed file and/or migrations there would be no need for the job and all associated code. We'd also be able to release the changes under a Standard RfC, which means the changes could appear as quickly as the next day.

They are now managed by a seed process in [water-abstraction-system](https://github.com/DEFRA/water-abstraction-system) so now is as good a time as any to get rid of it!