Remove job from startup registration
We're following a pattern we've adopted for dealing with the legacy code. We're not comfortable deleting everything related to a 'feature', because we are not confident we would know what 'everything' is!

So, we remove or disable features at the edges. Over time we monitor the logs and app activity, and as we see it know longer being used, we know a legacy app is ready to archive.