Actually stop charge categories sync job! (#2688)
In [Remove charge categories sync job](https://github.com/DEFRA/water-abstraction-service/pull/2687), we thought we'd stopped the charge categories sync job from running. As we highlighted in that change, we don't tend to remove code, because we are wary of how stuff is often reused in the legacy code, and deleting code can break things.

So, instead, we tend to remove things at the 'edges'. In #2687, we stopped starting the job when the app starts. We know it has config to automatically repeat the job, so we figured not starting it would mean it couldn't repeat. How wrong we were!

Looking at the logs in `pre`, it still seems to be running. So, in this change, we'll stop the job itself from registering and remove the repeat config.