Update returns for company endpoint for quarterly (#2692)
https://eaflood.atlassian.net/browse/WATER-4796

We are updating the service to allow customers to upload returns quarterly (four times a year instead of just once).

The current logic is tied to displaying the link and generating the CSVs for the cycle you are currently in. To enable quarterly, we have to break that tie-in and move to a more flexible "I have a due return; give me the CSV!"

We've already made a change to allow the link always to be displayed (if you have bulk uploads enabled and have due returns) in [Always display bulk upload link on external site](https://github.com/DEFRA/water-abstraction-ui/pull/2677).

This change supports our changes to the [CSV template generation](https://github.com/DEFRA/water-abstraction-ui/pull/2679). The logic in the UI calls the `/water/1.0/company/{entityId}/returns` endpoint to fetch the returns for the current 'company'.

The CSV template will be grouped by due date, but to do that, the endpoint needs to start returning it.