Remove date filter on internal search for returns
https://eaflood.atlassian.net/browse/WATER-5085

Whilst testing another feature of the service it was spotted that the way we are creating return requirements is not the same as NALD (NALD now uses unique references, we weren't!)

One of the things that highlighted this was multiple results appearing in the internal search. The users were used to only ever seeing a single result.

There _are_ return requirements with duplicate references that we have imported from NALD, but they only apply to much older ones (pre 2008-04-01). We couldn't understand why searches for these returned no results until we found this filter.

Perhaps, knowing that WRLS was not originally importing old return submission data there was felt to be no need to include them. But now _all_ NALD returns data has been imported into WRLS, and we are taking over the 'returns leg' from NALD we need to remove this restriction on the internal search.