Migration - update return version end dates (#2697)
https://eaflood.atlassian.net/browse/WATER-5099

This issue was found in production when users created new return versions for water companies that need quarterly returns enabled. When inserting a new return version between other existing return versions it was setting the end date of one of the return versions incorrectly. This was due to the order of the functions that determine which end date to use.

This has been fixed in DEFRA/water-abstraction-system#2095

This PR creates a migration to update impacted return versions so they will have the correct end date.