Include returns for 'ended' licences in search (#2704)
https://eaflood.atlassian.net/browse/WATER-5127

Whilst testing WATER-5006, our QA team raised a concern that something was broken.

They were searching for a return reference they could see in a licence's returns, but the search was saying 'No results found'.

When we investigated, we found the legacy search has been written to exclude results from licences that have ended. It will only search for return references for 'current licences'.

We don't know what value there was in this in the first place. However, now that WRLS is responsible for managing returns, this quirk is not only confusing but could also block users from carrying out their duties.